+++ lib/controllers/quiz/quiz_controller.dart
@@ -1,7 +1,7 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:iruka_quiz/controllers/quiz/quiz_state.dart';
 
-import '../../models/question.dart';
+import '../../models/question/question.dart';
 
 final quizControllerProvider =
     StateNotifierProvider.autoDispose<QuizController, QuizState>(
